Use four-terminal current sense resistor symbol
Previously we had used a special four-terminal symbol and footprint to
ensure that current sense signals did not connect to power supply zones
in multiple places. We rejected this in favor of a two-terminal
schematic symbol and standard resistor footprint to ensure reliable
assembly.

This new solution uses a four-terminal schematic symbol with a
two-terminal footprint which is actually four terminals in two
overlapping pairs using the "net tie" footprint feature.

Fixes #13