Change TARGET_C_VBUS pass-through default
Previously pass-through power between TARGET C and TARGET A was enabled
by default, even when the Cynthion was unpowered. This had the potential
benefit of allowing a pass-through target device to function while the
Cynthion was off, but the feature was unreliable due to loading of D+/D-
by the PHY and TARGET_FS_MONITOR connections. All pass-through power is
now disabled by default, providing a consistent user experience while
the Cynthion is unpowered.